Planning: move path_assessment_decider into a single folder.
