+++ modules/planning/tasks/task_factory.cc
@@ -27,7 +27,7 @@
 #include "modules/planning/tasks/deciders/open_space_fallback_decider.h"
 #include "modules/planning/tasks/deciders/open_space_pre_stop_decider.h"
 #include "modules/planning/tasks/deciders/open_space_roi_decider.h"
-#include "modules/planning/tasks/deciders/path_assessment_decider.h"
+#include "modules/planning/tasks/deciders/path_assessment_decider/path_assessment_decider.h"
 #include "modules/planning/tasks/deciders/path_bounds_decider/path_bounds_decider.h"
 #include "modules/planning/tasks/deciders/path_lane_borrow_decider/path_lane_borrow_decider.h"
 #include "modules/planning/tasks/deciders/speed_bounds_decider/speed_bounds_decider.h"
